+++ game/231/RB_Banner.c
@@ -84,13 +84,17 @@ int RB_Banner_Animate_Init(struct ModelHeader *mh)
 	u8 *vertex;
 	int count = 0;
 
-	if ((s16)(*(u16 *)(void *)mh->ptrCommandList) < 0x40)
+	// NOTE: ptrCommandList is a checkpoint-tracked fixed 4-byte "retail
+	// 32-bit RAM address slot" (see platform/native_checkpoint.c); round-trip
+	// through uintptr_t instead of casting the u32 value directly to a
+	// pointer, which truncates on 64-bit targets.
+	if ((s16)(*(u16 *)(void *)(uintptr_t)mh->ptrCommandList) < 0x40)
 	{
 		return 0;
 	}
 
 	vertex = RB_Banner_FirstVertex(mh);
-	cmd = (u32 *)((u8 *)mh->ptrCommandList + 4);
+	cmd = (u32 *)((u8 *)(uintptr_t)mh->ptrCommandList + 4);
 
 	while (*cmd != 0xffffffffU)
 	{
