+++ game/GhostTape.c
@@ -69,7 +69,10 @@ void GhostTape_End(void)
 	gh->ySpeed = d->ySpeed;
 	gh->speedApprox = d->speedApprox;
 	gh->timeElapsedInRace = d->timeElapsedInRace;
-	gh->size = (u32)sdata->GhostRecording.ptrCurrOffset - (u32)sdata->GhostRecording.ptrStartOffset;
+	// NOTE: (u32)ptr truncates real pointers on 64-bit (Switch/AArch64); use
+	// real pointer subtraction instead (identical result on 32-bit hosts,
+	// and correct regardless of absolute address on 64-bit).
+	gh->size = (u32)(sdata->GhostRecording.ptrCurrOffset - sdata->GhostRecording.ptrStartOffset);
 	NativeGhostInput_StopRecording();
 }
 
@@ -261,7 +264,10 @@ void GhostTape_WriteMoves(s16 raceFinished)
 		    // if offset of ghost-recording buffer exceeds
 		    // the maximum size of a ghost that can be recorded
 		    // (if you're one frame away from max capacity)
-		    ((u32)sdata->GhostRecording.ptrEndOffset < (u32)writeCursor + GHOST_RECORD_BUFFER_END_GUARD) &&
+		    // NOTE: (u32)ptr truncates real pointers on 64-bit
+		    // (Switch/AArch64); compare pointers directly instead (both
+		    // point within the same recording buffer).
+		    (sdata->GhostRecording.ptrEndOffset < writeCursor + GHOST_RECORD_BUFFER_END_GUARD) &&
 
 		    (sdata->boolCanSaveGhost = 0,
